#642 move out remaining files in original components folder
