+++ src/client/pages/ProjectsPage/projects.tsx
@@ -6,7 +6,7 @@
 import { Route, Switch } from 'react-router-dom';
 import { routes } from '../../../shared/routes';
 import ProjectsView from './projects-view';
-import WBSDetails from '../../../components/projects/wbs-details/wbs-details';
+import WBSDetails from '../Misc/wbs-details';
 import './projects.module.css';
 import CreateProjectForm from '../CreateProjectPage/create-project-form';
 import CreateWPForm from '../CreateWorkPackagePage/create-wp-form';
